Refactored Fisher features to use clean architecture and Cubits
This commit migrates the Fisher feature set from legacy BLoC implementations to a new Cubit-based Clean Architecture. It replaces direct repository usage with interface-based dependency injection, integrates Domain Entities and Value Objects, and cleans up legacy data models.

Key changes:
-   **State Management Migration:**
    -   Replaced `CatchesBloc`, `OffersBloc`, and `OrdersBloc` with `CatchesCubit`, `OffersCubit`, and `OrdersCubit`.
    -   Updated `main.dart` and `injector.dart` to register the new Cubits and bind repository interfaces (`ICatchRepository`, `IOfferRepository`, `IOrderRepository`) instead of concrete legacy implementations.
    -   Deleted legacy BLoC files and obsolete repositories (`CatchRepository`, `OfferRepository`, `OrderRepository` in `fisher/data`).

-   **Domain & Data Layer:**
    -   **Entities:** Deleted legacy models (`catch.dart`, `offer.dart`, `order.dart` in `core/models`) in favor of Domain Entities and Value Objects (e.g., `Weight`, `Price`, `OfferTerms`).
    -   **Repositories:** Updated `OrderRepositoryImpl` and `OfferRepositoryImpl` to return entities and throw domain exceptions (`NotFoundException`) instead of returning nullable models.
    -   **Seeder:** Refactored `seeder.dart` to utilize the new interfaces and instantiate Value Objects for demo data.

-   **UI Integration:**
    -   **Fisher Home:** Updated `FisherHome` to use `OrdersCubit`, `CatchesCubit`, and `OffersCubit`; refactored turnover calculation and "For Sale/Sold" tab logic.
    -   **Catch Details:** Migrated `CatchDetails` to load data via `CatchesCubit` and `OffersCubit`; updated edit/delete actions to use the new state methods.
    -   **Order Details:** Refactored `OrderDetails` to use `OrdersCubit` and `ICatchRepository`; mapped UI fields to `Order` entity properties (e.g., `terms.weight.grams`).
    -   **Offer Details:** Connected `FisherOfferDetails` to `OffersCubit`; implemented `NegotiationService` calls for accept, reject, and counter-offer actions; added view tracking (`markOfferAsViewed`).
    -   **Notifications:** Updated `NotificationsScreen` to consume `OffersCubit` for filtering pending offers.

-   **Utilities:**
    -   Added `generate_docs.py` for API specification documentation generation.

+++ lib/bloc/cubits/products_filter_cubit/products_filter_state.dart
@@ -3,7 +3,7 @@ part of 'products_filter_cubit.dart';
 class ProductsFilterState extends Equatable {
   final List<Species> selectedSpecies;
   final List<String> selectedLocations;
-  final double minWeight;
+  final int minWeight;
   final SortBy sortByPrice;
   final SortBy sortByDate;
   final bool applyFilters;
@@ -22,7 +22,7 @@ class ProductsFilterState extends Equatable {
   ProductsFilterState copyWith({
     List<Species>? selectedSpecies,
     List<String>? selectedLocations,
-    double? minWeight,
+    int? minWeight,
     SortBy? sortByPrice,
     SortBy? sortByDate,
     bool? applyFilters,
